Made it play a song when the Caps Lock changes state
The old behavior was playing a sound when the LED state changes
and the Caps Lock is on.  This was helpful in the sense that
it would beep as long as the Caps Lock was on.  But it was a little
annoying too.  Also, it only beeped if you hit some keys that would
change the LED state.

Now it beeps when the Caps Lock changes state only.  I think it's
enough information for the user.